Move types into types.ts
