--- code/lib/builder-vite/src/types/extended-options.type.ts
@@ -1,8 +1,0 @@
-import type { Options } from '@storybook/types';
-
-// Using instead of `Record<string, string>` to provide better aware of used options
-type IframeOptions = {
-  title: string;
-};
-
-export type ExtendedOptions = Options & IframeOptions;
